revamp ASB connector
