+++ asb-ballerina/Package.md
@@ -14,7 +14,7 @@ This package provides the capability to easily access Microsoft Azure Service Bu
 
 ## Report issues
 
-To report bugs, request new features, start new discussions, view project boards, etc., go to the [Ballerina Microsoft Azure Service Bus repository](https://github.com/ballerina-platform/module-ballerinax-azure-service-bus)
+To report bugs, request new features, start new discussions, view project boards, etc., go to the [Ballerina Extended Library repository](https://github.com/ballerina-platform/ballerina-extended-library)
 
 ## Useful links
 - Discuss code changes of the Ballerina project in [ballerina-dev@googlegroups.com](mailto:ballerina-dev@googlegroups.com).
