+++ asb-ballerina/samples/deadletter_from_queue.bal
@@ -45,23 +45,21 @@ public function main() returns error? {
         connectionString: connectionString
     };
 
-    asb:AsbClient asbClient = new (config);
+    log:printInfo("Initializing Asb sender client.");
+    asb:MessageSender queueSender = check new(connectionString, queueName);
 
-    log:printInfo("Creating Asb sender connection.");
-    handle queueSender = check asbClient->createQueueSender(queueName);
+    log:printInfo("Initializing Asb receiver client.");
+    asb:MessageReceiver queueReceiver = check new(connectionString, queueName, asb:PEEKLOCK);
 
-    log:printInfo("Creating Asb receiver connection.");
-    handle queueReceiver = check asbClient->createQueueReceiver(queueName, asb:PEEKLOCK);
+    log:printInfo("Sending via Asb sender client.");
+    check queueSender->send(message1);
 
-    log:printInfo("Sending via Asb sender connection.");
-    check asbClient->send(queueSender, message1);
-
-    log:printInfo("Receiving from Asb receiver connection.");
-    asb:Message|asb:Error? messageReceived = asbClient->receive(queueReceiver, serverWaitTime);
+    log:printInfo("Receiving from Asb receiver client.");
+    asb:Message|asb:Error? messageReceived = queueReceiver->receive(serverWaitTime);
 
     if (messageReceived is asb:Message) {
-        check asbClient->deadLetter(queueReceiver, messageReceived);
-        asb:Message|asb:Error? messageReceivedAgain = asbClient->receive(queueReceiver, serverWaitTime);
+        check queueReceiver->deadLetter(messageReceived);
+        asb:Message|asb:Error? messageReceivedAgain = queueReceiver->receive(serverWaitTime);
         if (messageReceivedAgain is ()) {
             log:printInfo("Deadletter message successful");
         } else {
@@ -73,9 +71,9 @@ public function main() returns error? {
         log:printError("Receiving message via Asb receiver connection failed.");
     }
 
-    log:printInfo("Closing Asb sender connection.");
-    check asbClient->closeSender(queueSender);
+    log:printInfo("Closing Asb sender client.");
+    check queueSender->close();
 
-    log:printInfo("Closing Asb receiver connection.");
-    check asbClient->closeReceiver(queueReceiver);
+    log:printInfo("Closing Asb receiver client.");
+    check queueReceiver->close();
 }
